Finally figured out how to get rid of the annoying custom-variables
custom-variables set by emacs was clogging my init.el and my git
history. This offloads them to a temp file. It is important that it
be set near immediately, so make sure that the custom file set is at
the top of editing.el and editing.el is loaded first or second